+++ tscode/operators.py
@@ -18,6 +18,7 @@ GNU General Public License for more details.
 
 import os
 import pickle
+import pickle
 import time
 from subprocess import DEVNULL, STDOUT, check_call
 
@@ -25,13 +26,18 @@ import numpy as np
 from networkx import connected_components
 
 from tscode.clustered_csearch import csearch
+from tscode.clustered_csearch import csearch
 from tscode.errors import InputError
 from tscode.graph_manipulations import graphize
 from tscode.hypermolecule_class import align_structures
 from tscode.optimization_methods import _refine_structures, optimize
 from tscode.pka import pka_routine
+from tscode.hypermolecule_class import align_structures
+from tscode.optimization_methods import _refine_structures, optimize
+from tscode.pka import pka_routine
 from tscode.settings import (CALCULATOR, DEFAULT_FF_LEVELS, DEFAULT_LEVELS,
                              FF_CALC, FF_OPT_BOOL, PROCS)
+from tscode.utils import (get_scan_peak_index, read_xyz,
 from tscode.utils import (get_scan_peak_index, read_xyz,
                           suppress_stdout_stderr, time_to_string, write_xyz)
 from tscode.ase_manipulations import ase_saddle
@@ -102,6 +108,14 @@ def operate(input_string, embedder):
         op = input_string.split('>')[0]
         raise Exception(f'Operator {op} not recognized.')
 
+    elif 'pka>' in input_string:
+        pka_routine(filename, embedder)
+        outname = filename
+
+    else:
+        op = input_string.split('>')[0]
+        raise Exception(f'Operator {op} not recognized.')
+
     return outname
 
 def confab_operator(filename, options, logfunction=None):
@@ -139,6 +153,9 @@ def confab_operator(filename, options, logfunction=None):
     # if len(conformers) > 10 and not options.let:
     #     conformers = conformers[0:10]
     #     logfunction(f'Will use only the best 10 conformers for TSCoDe embed.\n')
+    # if len(conformers) > 10 and not options.let:
+    #     conformers = conformers[0:10]
+    #     logfunction(f'Will use only the best 10 conformers for TSCoDe embed.\n')
 
     os.remove(confname)
     with open(confname, 'w') as f:
@@ -148,6 +165,7 @@ def confab_operator(filename, options, logfunction=None):
     return confname
 
 def csearch_operator(filename, embedder, keep_hb=False, mode=1):
+def csearch_operator(filename, embedder, keep_hb=False, mode=1):
     '''
     '''
 
@@ -155,7 +173,12 @@ def csearch_operator(filename, embedder, keep_hb=False, mode=1):
     if keep_hb:
         s += ' (preserving current hydrogen bonds)'
     embedder.log(s)
+    s = f'--> Performing conformational search on {filename}'
+    if keep_hb:
+        s += ' (preserving current hydrogen bonds)'
+    embedder.log(s)
 
+    # t_start = time.perf_counter()
     # t_start = time.perf_counter()
 
     data = read_xyz(filename)
@@ -165,10 +188,13 @@ def csearch_operator(filename, embedder, keep_hb=False, mode=1):
                       'an individual search from each conformer (might be time-consuming).')
                                 
     # calc, method, procs = _get_lowest_calc(embedder)
+    # calc, method, procs = _get_lowest_calc(embedder)
     conformers = []
 
     for i, coords in enumerate(data.atomcoords):
 
+        # opt_coords = optimize(coords, data.atomnos, calculator=calc, method=method, procs=procs)[0] if embedder.options.optimization else coords
+        opt_coords = coords
         # opt_coords = optimize(coords, data.atomnos, calculator=calc, method=method, procs=procs)[0] if embedder.options.optimization else coords
         opt_coords = coords
         # optimize starting structure before running csearch
@@ -184,10 +210,24 @@ def csearch_operator(filename, embedder, keep_hb=False, mode=1):
                                 logfunction=embedder.log,
                                 write_torsions=embedder.options.debug
                             )
+        conf_batch = csearch(
+                                opt_coords,
+                                data.atomnos,
+                                constrained_indexes=_get_internal_constraints(filename, embedder),
+                                keep_hb=keep_hb,
+                                mode=mode,
+                                n_out=embedder.options.max_confs,
+                                title=f'{filename}_conf{i}',
+                                logfunction=embedder.log,
+                                write_torsions=embedder.options.debug
+                            )
         # generate the most diverse conformers starting from optimized geometry
 
         conformers.extend(conf_batch)
+        conformers.extend(conf_batch)
 
+    conformers = np.concatenate(conformers)
+    # batch_size = conformers.shape[1]
     conformers = np.concatenate(conformers)
     # batch_size = conformers.shape[1]
 
@@ -199,11 +239,18 @@ def csearch_operator(filename, embedder, keep_hb=False, mode=1):
     #     conformers = most_diverse_conformers(batch_size, conformers)
 
     print(f'Writing conformers to file...{" "*10}', end='\r')
+    # if embedder.embed is not None:
+    #     embedder.log(f'\nSelected the most diverse {batch_size} out of {conformers.shape[0]} conformers for {filename} ({time_to_string(time.perf_counter()-t_start)})')
+    #     conformers = most_diverse_conformers(batch_size, conformers)
+
+    print(f'Writing conformers to file...{" "*10}', end='\r')
 
     confname = filename[:-4] + '_confs.xyz'
     with open(confname, 'w') as f:
         for i, conformer in enumerate(conformers):
             write_xyz(conformer, data.atomnos, f, title=f'Generated conformer {i}')
+
+    print(f'{" "*30}', end='\r')
 
     print(f'{" "*30}', end='\r')
 
@@ -222,12 +269,21 @@ def opt_operator(filename, embedder, logfunction=None):
     mol = next((mol for mol in embedder.objects if mol.name == filename))
     # load molecule to be optimized from embedder
 
+    mol = next((mol for mol in embedder.objects if mol.name == filename))
+    # load molecule to be optimized from embedder
+
     if logfunction is not None:
+        logfunction(f'--> Performing {embedder.options.calculator} {embedder.options.theory_level}' + (
+                    f'{f"/{embedder.options.solvent}" if embedder.options.solvent is not None else ""} optimization on {filename} ({len(mol.atomcoords)} conformers)'))
         logfunction(f'--> Performing {embedder.options.calculator} {embedder.options.theory_level}' + (
                     f'{f"/{embedder.options.solvent}" if embedder.options.solvent is not None else ""} optimization on {filename} ({len(mol.atomcoords)} conformers)'))
                                 
     energies = []
     lowest_calc = _get_lowest_calc(embedder)
+
+    t_start = time.perf_counter()
+
+    conformers, energies = _refine_structures(mol.atomcoords, mol.atomnos, *lowest_calc, loadstring='Optimizing conformer', logfunction=lambda s:embedder.log(s, p=False))
 
     t_start = time.perf_counter()
 
@@ -240,11 +296,16 @@ def opt_operator(filename, embedder, logfunction=None):
 
     mask = energies < 20
     # getting the structures to reject (Rel Energy > 20 kcal/mol)
+    mask = energies < 20
+    # getting the structures to reject (Rel Energy > 20 kcal/mol)
 
     if logfunction is not None:
         s = 's' if len(conformers) > 1 else ''
         s = f'Completed optimization on {len(conformers)} conformer{s}. ({time_to_string(time.perf_counter()-t_start)}, ~{time_to_string((time.perf_counter()-t_start)/len(conformers))} per structure).\n'
+        s = f'Completed optimization on {len(conformers)} conformer{s}. ({time_to_string(time.perf_counter()-t_start)}, ~{time_to_string((time.perf_counter()-t_start)/len(conformers))} per structure).\n'
 
+        if max(energies) > 20:
+            s += f'Discarded {len(conformers)-np.count_nonzero(mask)}/{len(conformers)} unstable conformers (Rel. E. > 20 kcal/mol)\n'
         if max(energies) > 20:
             s += f'Discarded {len(conformers)-np.count_nonzero(mask)}/{len(conformers)} unstable conformers (Rel. E. > 20 kcal/mol)\n'
 
@@ -255,6 +316,11 @@ def opt_operator(filename, embedder, logfunction=None):
     with open(optname, 'w') as f:
         for i, conformer in enumerate(align_structures(conformers)):
             write_xyz(conformer, mol.atomnos, f, title=f'Optimized conformer {i} - Rel. E. = {round(energies[i], 3)} kcal/mol')
+        for i, conformer in enumerate(align_structures(conformers)):
+            write_xyz(conformer, mol.atomnos, f, title=f'Optimized conformer {i} - Rel. E. = {round(energies[i], 3)} kcal/mol')
+
+    logfunction(s+'\n')
+    logfunction(f'Wrote {len(conformers)} optimized structures to {optname}\n')
 
     logfunction(s+'\n')
     logfunction(f'Wrote {len(conformers)} optimized structures to {optname}\n')
@@ -267,6 +333,8 @@ def neb_operator(filename, embedder, attempts=5):
     '''
     embedder.t_start_run = time.perf_counter()
     data = read_xyz(filename)
+    n_str = len(data.atomcoords)
+    assert (n_str in (2, 3) or n_str % 2 == 1), 'NEB calculations need a .xyz input file with two, three or an odd number of geometries.'
     n_str = len(data.atomcoords)
     assert (n_str in (2, 3) or n_str % 2 == 1), 'NEB calculations need a .xyz input file with two, three or an odd number of geometries.'
 
@@ -329,7 +397,11 @@ def neb_operator(filename, embedder, attempts=5):
 
         print('Getting start point energy...', end='\r')
         _, reag_energy, _ = ase_popt(embedder, reagents, data.atomnos, steps=0)
+        print('Getting start point energy...', end='\r')
+        _, reag_energy, _ = ase_popt(embedder, reagents, data.atomnos, steps=0)
 
+        print('Getting end point energy...', end='\r')
+        _, prod_energy, _ = ase_popt(embedder, products, data.atomnos, steps=0)
         print('Getting end point energy...', end='\r')
         _, prod_energy, _ = ase_popt(embedder, products, data.atomnos, steps=0)
 
@@ -376,6 +448,7 @@ def neb_operator(filename, embedder, attempts=5):
         embedder.log(f'\nNEB failed, TS energy is lower than both the start and end points.\n')
 
     with open(f'{title}_TS.xyz', 'w') as f:
+    with open(f'{title}_TS.xyz', 'w') as f:
         write_xyz(ts_coords, data.atomnos, f, title='NEB TS - see log for relative energies')
 
 def saddle_operator(filename, embedder):
@@ -418,8 +491,12 @@ def scan_operator(filename, embedder):
     '''
     Thought to approach or separate two reactive atoms, looking for the energy maximum.
     Scan direction is inferred by the reactive index distance.
+    Thought to approach or separate two reactive atoms, looking for the energy maximum.
+    Scan direction is inferred by the reactive index distance.
     '''
     embedder.t_start_run = time.perf_counter()
+    mol_index, mol = next(((i, mol) for i, mol in enumerate(embedder.objects) if mol.name == filename))
+
     mol_index, mol = next(((i, mol) for i, mol in enumerate(embedder.objects) if mol.name == filename))
 
     assert len(mol.atomcoords) == 1, 'The scan> operator works on a single .xyz geometry.'
@@ -435,8 +512,18 @@ def scan_operator(filename, embedder):
     t_start = time.perf_counter()
 
     # shorthands for clearer code
+    t_start = time.perf_counter()
+
+    # shorthands for clearer code
     i1, i2 = mol.reactive_indexes
     coords = mol.atomcoords[0]
+
+    # getting the start distance between scan indexes and start energy
+    d = norm_of(coords[i1]-coords[i2])
+
+    # deciding if moving atoms closer or further apart based on distance
+    bonds = list(graphize(coords, mol.atomnos).edges)
+    step = 0.05 if (i1, i2) in bonds else -0.05
 
     # getting the start distance between scan indexes and start energy
     d = norm_of(coords[i1]-coords[i2])
@@ -448,12 +535,17 @@ def scan_operator(filename, embedder):
     # logging to file and terminal
     embedder.log(f'--> {mol.rootname} - Performing a distance scan {"approaching" if step < 0 else "separating"} indexes {i1} ' +
                  f'and {i2} - step size {round(step, 2)} A\n    Theory level is {embedder.options.theory_level} ' +
+    # logging to file and terminal
+    embedder.log(f'--> {mol.rootname} - Performing a distance scan {"approaching" if step < 0 else "separating"} indexes {i1} ' +
+                 f'and {i2} - step size {round(step, 2)} A\n    Theory level is {embedder.options.theory_level} ' +
                  f'via {embedder.options.calculator}')
 
     # creating a dictionary that will hold results
     # and the structure output list
     dists, energies, structures = [], [], []
+    dists, energies, structures = [], [], []
 
+    # getting atomic symbols
     # getting atomic symbols
     s1, s2 = mol.atomnos[[i1, i2]]
 
@@ -523,6 +615,7 @@ def scan_operator(filename, embedder):
     ### Start the plotting sequence
 
     fig = plt.figure()
+    fig = plt.figure()
     plt.plot(
         dists,
         energies,
@@ -548,6 +641,7 @@ def scan_operator(filename, embedder):
     )
 
     title = mol.rootname + ' distance scan'
+    title = mol.rootname + ' distance scan'
     plt.legend()
     plt.title(title)
     plt.xlabel(f'Indexes {i1}-{i2} distance (A)')
@@ -558,15 +652,18 @@ def scan_operator(filename, embedder):
     plt.ylabel('Rel. E. (kcal/mol)')
     plt.savefig(f'{title.replace(" ", "_")}_plt.svg')
     pickle.dump(fig, open(f'{title.replace(" ", "_")}_plt.pickle', 'wb'))
+    pickle.dump(fig, open(f'{title.replace(" ", "_")}_plt.pickle', 'wb'))
 
     ### Start structure writing 
 
+    # print all scan structures
     # print all scan structures
     with open(f'{mol.name[:-4]}_scan.xyz', 'w') as f:
         for i, (s, d, e) in enumerate(zip(structures, dists, energies)):
             write_xyz(s, mol.atomnos, f, title=f'Scan point {i+1}/{len(structures)} ' +
                       f'- d({i1}-{i2}) = {round(d, 2)} A - Rel. E = {round(e, 2)} kcal/mol')
 
+    # print the maximum on another file for convienience
     # print the maximum on another file for convienience
     with open(f'{mol.name[:-4]}_scan_max.xyz', 'w') as f:
         s = structures[id_max]
@@ -576,7 +673,11 @@ def scan_operator(filename, embedder):
 
     embedder.log(f'\n--> Written {len(structures)} structures to {mol.name[:-4]}_scan.xyz ({time_to_string(time.perf_counter() - t_start)})')
     embedder.log(f'\n--> Written energy maximum to {mol.name[:-4]}_scan_max.xyz\n')
+    embedder.log(f'\n--> Written {len(structures)} structures to {mol.name[:-4]}_scan.xyz ({time_to_string(time.perf_counter() - t_start)})')
+    embedder.log(f'\n--> Written energy maximum to {mol.name[:-4]}_scan_max.xyz\n')
 
+    # Log data to the embedder class
+    embedder.objects[mol_index].scan_data = (dists, energies)
     # Log data to the embedder class
     embedder.objects[mol_index].scan_data = (dists, energies)
 
@@ -589,11 +690,27 @@ def _get_lowest_calc(embedder=None):
     if embedder is None:
         if FF_OPT_BOOL:
             return (FF_CALC, DEFAULT_FF_LEVELS[FF_CALC], PROCS)
+            return (FF_CALC, DEFAULT_FF_LEVELS[FF_CALC], PROCS)
         return (CALCULATOR, DEFAULT_LEVELS[CALCULATOR], PROCS)
 
     if embedder.options.ff_opt:
         return (embedder.options.ff_calc, embedder.options.ff_level, embedder.options.procs)
+        return (embedder.options.ff_calc, embedder.options.ff_level, embedder.options.procs)
     return (embedder.options.calculator, embedder.options.theory_level, embedder.options.procs)
+
+def _get_internal_constraints(filename, embedder):
+    '''
+    '''
+    mol_id = next((i for i, mol in enumerate(embedder.objects) if mol.name == filename))
+    # get embedder,objects index of molecule to get internal constraints of
+
+    out = []
+    for _, tgt in embedder.pairings_dict[mol_id].items():
+        if isinstance(tgt, tuple):
+            out.append(tgt)
+
+    return np.array(out)
+
 
 def _get_internal_constraints(filename, embedder):
     '''